# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: List the current user's withdrawals
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request a new withdrawal (enters pending state)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "200":
          description: Pending withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: List the current user's deposits
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Record a new deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Created deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
  /moonpay/initiate:
    post:
      operationId: initiateMoonpayBuy
      tags: [moonpay]
      summary: Build a MoonPay hosted-checkout URL for a Buy Crypto flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoonpayInitiateRequest"
      responses:
        "200":
          description: MoonPay checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoonpayInitiateResponse"
...
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, awaiting_gas_fee, approved, rejected, completed, cancelled, expired]
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer]
        destination:
          type: string
        status:
          type: string
          enum: [pending, awaiting_gas_fee, approved, rejected, completed, cancelled, expired]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        decidedAt:
          type: string
          nullable: true
        gasFeeAmount:
          type: number
          nullable: true
          description: ETH gas fee amount the admin set on this withdrawal that the user must fund.
        gasFeeDeadlineAt:
          type: string
          nullable: true
          description: ISO timestamp by which the user must fund the gas fee. After this, the withdrawal expires.
        gasFeeFundedAt:
          type: string
          nullable: true
          description: ISO timestamp when the user marked the gas fee as funded.
        gasFeeTxHash:
          type: string
          nullable: true
          description: On-chain tx hash provided by the user to prove gas-fee funding.
        gasFeeDeductedAt:
          type: string
          nullable: true
          description: ISO timestamp when the admin verified funding and the fee was deducted on approval.
      required: [id, userId, userName, amount, currency, method, destination, status, createdAt]
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.000001
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer]
        destination:
          type: string
        sourceWalletId:
          type: string
          nullable: true
          description: |
            When supplied, the withdrawal is funded by directly broadcasting
            an on-chain transfer from this connected external wallet to the
            destination address. The server signs and broadcasts via
            ethers.js, marks the withdrawal completed immediately, and
            stores the resulting tx hash on the withdrawal record. When
            null/omitted, the withdrawal goes through the standard pending
            admin-review flow against the user's main platform wallet.
      required: [amount, currency, method, destination]
...
    Deposit:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer, card]
        status:
          type: string
          enum: [pending, completed, failed]
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, amount, currency, method, status, createdAt]
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.000001
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer, card]
        reference:
          type: string
          nullable: true
        externalWalletId:
          type: string
          nullable: true
          description: When method is crypto_wallet and the funds are sent from a connected external wallet, the id of that wallet.
        txHash:
          type: string
          nullable: true
          description: On-chain transaction hash recorded after a successful external_wallet send.
      required: [amount, currency, method]
...
    MoonpayInitiateRequest:
      type: object
      properties:
        assetSymbol:
          type: string
          description: Crypto asset symbol to purchase (e.g. BTC, ETH, USDT).
        fiatAmount:
          type: number
          minimum: 1
          description: Fiat amount the user wants to spend.
        fiatCurrency:
          type: string
          description: Fiat currency code (e.g. USD, EUR).
        destinationAddress:
          type: string
          description: Wallet address that should receive the purchased crypto.
        destinationKind:
          type: string
          enum: [platform, external, custom]
          description: Source of the destination address.
      required: [assetSymbol, fiatAmount, fiatCurrency, destinationAddress, destinationKind]
    MoonpayInitiateResponse:
      type: object
      properties:
        url:
          type: string
          description: Hosted MoonPay checkout URL the user should be redirected to.
        sandbox:
          type: boolean
          description: True when the URL points at MoonPay sandbox (no live API key configured).
        signed:
          type: boolean
          description: True when the URL was HMAC-signed using the configured MoonPay secret.
        configured:
          type: boolean
          description: True when MOONPAY_API_KEY is configured server-side.
        autoFilled:
          type: boolean
          description: True when user identity (email/name) was auto-filled into the URL.
        notice:
          type: string
          nullable: true
          description: Optional human-readable notice to surface in the dialog.
      required: [url, sandbox, signed, configured, autoFilled]
...